Rename "files" directory to "resources"
We rename this directory since we have modules that actually deal with files/blobs, but "file(s)" is already taken up by this. "files" is very general term, while "resources" more precisely describe the role of the files in the directory.
