+++ WebRandomizer/ClientApp/src/resources/markdown/instructions.md
