+++ WebRandomizer/ClientApp/src/resources/sprite/inventory.json
